[FIX] Lint (#2030)
